backend/requirements.txt
# Processamento de Imagens
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# API REST
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilidades
python-dotenv>=1.0.0
aiofiles>=23.2.1
